requirements.txt
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.28
protobuf==3.20.1
PyJWT==2.3.0
PyMySQL==1.0.2
six==1.16.0
SQLAlchemy==1.4.35
Werkzeug==2.1.1
